\name{client_behavior}
\docType{data}
\alias{client_behavior}
\title{Simulated Client Behavior Data}
\description{
  This dataset simulates client characteristics and outcomes such as satisfaction and churn.
}
\usage{data(client_behavior)}
\format{
  A data frame with 1000 observations and 6 variables:
  \describe{
    \item{client_id}{Client identifier}
    \item{age}{Client age}
    \item{income}{Annual income}
    \item{satisfaction_score}{Score from 1 to 10}
    \item{treatment}{Indicator for advisory treatment}
    \item{churned}{Whether the client churned (1) or not (0)}
  }
}
