\name{macro_shocks}
\docType{data}
\alias{macro_shocks}
\title{Simulated Macroeconomic Shocks Time Series}
\description{
  A time-series dataset of macroeconomic variables affecting institutional returns.
}
\usage{data(macro_shocks)}
\format{
  A data frame with 120 monthly observations:
  \describe{
    \item{date}{Monthly observation date}
    \item{interest_rate}{Simulated interest rate}
    \item{gdp_growth}{GDP growth rate}
    \item{market_index}{Cumulative index for market performance}
  }
}
